backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
psycopg2-binary
aioredis
python-multipart
jose[cryptography]
passlib[bcrypt]
httpx
